Catering Backend

+++ src/main/java/za/ac/cput/repository/eventrepositories/ICateringRepository.java
@@ -3,5 +3,5 @@ package za.ac.cput.repository.eventrepositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import za.ac.cput.domain.eventdomains.Catering;
 
-public interface ICateringRepository extends JpaRepository<Catering, Integer> {
+public interface ICateringRepository extends JpaRepository<Catering, Long> {
 }
